function/requirements.txt
requests==2.28.2
google-api-core==2.11.0
google-cloud-bigquery==3.4.2
google-cloud-pubsub==2.14.0